Extract AtomicFile and ShowAsOwnedModal into shared helpers
Both were private implementations that a second caller now needs.

WriteAtomicAsync moves out of SettingsService into Helpers/AtomicFile so
anything persisting a file under %LocalAppData%\XrayUI gets the same
write-to-temp + replace guarantee.

CustomRulesWindow's hand-rolled owner/modal sequence moves into
Helpers/ModalWindowHelper as this.ShowAsOwnedModal(owner). The order is
load-bearing in ways the API does not signal and getting it wrong fails
silently, so it is worth having in one place rather than copied.

No behavior change.
